+++ src/lib/charts/createInvestmentBreakdownChart.js
