+++ src/lib/charts/createProductionChart.js
